Reorganize client-side code (WIP)
The ultimate goal is to adopt the container/component architecture.

For now we'll just put all legacy views into the 'components' folder, and split
each of them into pure component and container later.
