+++ imports/ui/components/pages/AddLine.js
@@ -15,7 +15,7 @@ import RadioGroup from '@material-ui/core/RadioGroup';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import FormControl from '@material-ui/core/FormControl';
 import FormLabel from '@material-ui/core/FormLabel';
-import MainLayout from './MainLayout';
+import MainLayout from '../lib/MainLayout';
 
 const AddLine = ({ history }) => {
   const [name, setName] = useState('');
