+++ imports/ui/components/pages/Feedback.js
@@ -10,8 +10,8 @@ import {
 } from 'react-onsenui';
 import { toast } from 'react-toastify';
 import { withTracker } from 'meteor/react-meteor-data';
-import { Locations } from '../api/lines.js';
-import MainLayout from './MainLayout';
+import { Locations } from '../../../api/lines.js';
+import MainLayout from '../lib/MainLayout';
 
 export default function feedback({}) {
   return (
